+++ odoo/local-src/specific_product/models/product_template.py
@@ -3,7 +3,6 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 import odoo.addons.decimal_precision as dp
-
 from odoo import fields, models