+++ odoo/local-src/specific_purchase/models/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+# © 2017 Okia SPRL
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+from . import res_partner
+from . import product_template
+from . import purchase_order
+from . import stock_picking
+from . import bank_holiday
+from . import purchase_config_settings