+++ HISTORY.rst
@@ -9,6 +9,7 @@ Unreleased
 **Data Migration**
 
 **Features and Improvements**
+* ALCN-823: Add the module "specific_purchase"
 * ALCN-831: Version 2 of the invoice report
 * ALCN-832: Import new account type
 * ALCN-833: Update chart of accounts (add account 0 and 8)